README: Add ToDo item for busy polling
Polling requires configuration of defer_hard_irqs and gro_flush_timeout.

With Linux v6.13 it is possible to configure this per NAPI instance
rather than global per network interface. This is extremely useful to run
real and non-real time traffic on the same network interface.

The idea is to add code to xdp which uses netlink to configure this.

Signed-off-by: Kurt Kanzenbach <kurt@linutronix.de>
Reviewed-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>