fix(specfile): correct path for checkout
There was an extra letter in the git checkout command that was making SPEC
file to failure in compilation when using rpmbuild by introducing an extra
'v' for the version we're checking out in the repo.

When building the RPM and SRPM packages with:
|rpmbuild -ba red\ hat/rtc-testbench.spec --define='version v5.2' --define='myrepo Linutronix/RTC-Testbench'

The command is properly substituting 'version' with 'v5.2' so git checkout
v{version} was adding an extra v.

Signed-off-by: Pablo Iranzo Gómez <Pablo.Iranzo@gmail.com>
Closes: https://github.com/Linutronix/TSN-Testbench/pull/26
[Kurt: Extend commit message]
Signed-off-by: Kurt Kanzenbach <kurt@linutronix.de>