stat: Add possibility to perform stat reset
Global statistics are collected until the application run ends. The
interesting part for real time is the average and worst case time over the
whole run. That is why the global statistics are not resetable in any way.

However, there are scanerios where this makes sense though. For example:

 - Start Testbench
 - Keep it runnung
 - Configure TSN switches, testing stuff in the network
 - Reset stats if valid configuration is found
 - => Avoid restart of Testbench

Therefore, add the possibility to reset the global statistics. Use SIGUSR1
signal for that.

Suggested-by: Ralf Kihm <Ralf.Kihm@kontron.com>
Signed-off-by: Kurt Kanzenbach <kurt@linutronix.de>
Reviewed-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>